requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Login==0.5.0
oauthlib==3.1.0
requests==2.23.0
pyOpenSSL==19.1.0
